refactor view test file
